+++ app/jobs/payment_providers/gocardless/handle_event_job.rb
@@ -5,9 +5,17 @@ module PaymentProviders
     class HandleEventJob < ApplicationJob
       queue_as 'providers'
 
-      def perform(events_json:)
-        result = PaymentProviders::GocardlessService.new.handle_event(events_json:)
-        result.raise_if_error!
+      def perform(organization: nil, events_json: nil, event_json: nil)
+        # NOTE: temporary keeps both events_json and event_json to avoid errors during the deployment
+        if events_json.present?
+          JSON.parse(events_json)['events'].each do |event|
+            PaymentProviders::Gocardless::HandleEventJob.perform_later(event_json: event.to_json)
+          end
+
+          return
+        end
+
+        PaymentProviders::Gocardless::HandleEventService.call(event_json:).raise_if_error!
       end
     end
   end
