+++ app/services/payment_providers/gocardless_service.rb
@@ -3,13 +3,6 @@
 module PaymentProviders
   class GocardlessService < BaseService
     REDIRECT_URI = "#{ENV["LAGO_OAUTH_PROXY_URL"]}/gocardless/callback".freeze
-    PAYMENT_ACTIONS = %w[paid_out failed cancelled customer_approval_denied charged_back].freeze
-    REFUND_ACTIONS = %w[created funds_returned paid refund_settled failed].freeze
-
-    PAYMENT_SERVICE_CLASS_MAP = {
-      "Invoice" => Invoices::Payments::GocardlessService,
-      "PaymentRequest" => PaymentRequests::Payments::GocardlessService
-    }.freeze
 
     def create_or_update(**args)
       access_token = if args[:access_code].present?
@@ -53,70 +46,6 @@ module PaymentProviders
       result.service_failure!(code: 'internal_error', message: e.description)
     end
 
-    def handle_incoming_webhook(organization_id:, body:, signature:, code: nil)
-      payment_provider_result = PaymentProviders::FindService.call(
-        organization_id:,
-        code:,
-        payment_provider_type: 'gocardless'
-      )
-
-      return payment_provider_result unless payment_provider_result.success?
-
-      events = GoCardlessPro::Webhook.parse(
-        request_body: body,
-        signature_header: signature,
-        webhook_endpoint_secret: payment_provider_result.payment_provider&.webhook_secret
-      )
-
-      PaymentProviders::Gocardless::HandleEventJob.perform_later(events_json: body)
-
-      result.events = events
-      result
-    rescue JSON::ParserError
-      result.service_failure!(code: 'webhook_error', message: 'Invalid payload')
-    rescue GoCardlessPro::Webhook::InvalidSignatureError
-      result.service_failure!(code: 'webhook_error', message: 'Invalid signature')
-    end
-
-    def handle_event(events_json:)
-      handled_events = []
-      events = JSON.parse(events_json)['events']
-      parsed_events = events.map { |event| GoCardlessPro::Resources::Event.new(event) }
-      parsed_events.each do |event|
-        case event.resource_type
-        when 'payments'
-          if PAYMENT_ACTIONS.include?(event.action)
-            update_payment_status_result = payment_service_klass(event)
-              .new.update_payment_status(
-                provider_payment_id: event.links.payment,
-                status: event.action
-              )
-
-            return update_payment_status_result unless update_payment_status_result.success?
-
-            handled_events << event
-          end
-        when 'refunds'
-          if REFUND_ACTIONS.include?(event.action)
-            status_result = CreditNotes::Refunds::GocardlessService
-              .new.update_status(
-                provider_refund_id: event.links.refund,
-                status: event.action,
-                metadata: event.metadata
-              )
-
-            return status_result unless status_result.success?
-
-            handled_events << event
-          end
-
-        end
-      end
-
-      result.handled_events = handled_events
-      result
-    end
-
     private
 
     def oauth
@@ -128,14 +57,6 @@ module PaymentProviders
         token_url: '/oauth/access_token',
         auth_scheme: :request_body
       )
-    end
-
-    def payment_service_klass(event)
-      payable_type = event.metadata["lago_payable_type"] || "Invoice"
-
-      PAYMENT_SERVICE_CLASS_MAP.fetch(payable_type) do
-        raise NameError, "Invalid lago_payable_type: #{payable_type}"
-      end
     end
   end
 end
