+++ spec/services/payment_providers/gocardless_service_spec.rb
@@ -118,156 +118,4 @@ RSpec.describe PaymentProviders::GocardlessService, type: :service do
       end
     end
   end
-
-  describe '.handle_incoming_webhook' do
-    let(:gocardless_provider) { create(:gocardless_provider, organization:) }
-    let(:events_result) { events['events'].map { |event| GoCardlessPro::Resources::Event.new(event) } }
-
-    let(:events) do
-      path = Rails.root.join('spec/fixtures/gocardless/events.json')
-      JSON.parse(File.read(path))
-    end
-
-    before { gocardless_provider }
-
-    it 'checks the webhook' do
-      allow(GoCardlessPro::Webhook).to receive(:parse)
-        .and_return(events_result)
-
-      result = gocardless_service.handle_incoming_webhook(
-        organization_id: organization.id,
-        body: events.to_json,
-        signature: 'signature'
-      )
-
-      expect(result).to be_success
-
-      expect(result.events).to eq(events_result)
-      expect(PaymentProviders::Gocardless::HandleEventJob).to have_been_enqueued
-    end
-
-    context 'when failing to parse payload' do
-      it 'returns an error' do
-        allow(GoCardlessPro::Webhook).to receive(:parse).and_raise(JSON::ParserError)
-
-        result = gocardless_service.handle_incoming_webhook(
-          organization_id: organization.id,
-          body: events.to_json,
-          signature: 'signature'
-        )
-
-        aggregate_failures do
-          expect(result).not_to be_success
-          expect(result.error).to be_a(BaseService::ServiceFailure)
-          expect(result.error.code).to eq('webhook_error')
-          expect(result.error.error_message).to eq('Invalid payload')
-        end
-      end
-    end
-
-    context 'when failing to validate the signature' do
-      it 'returns an error' do
-        allow(GoCardlessPro::Webhook).to receive(:parse)
-          .and_raise(GoCardlessPro::Webhook::InvalidSignatureError.new('error'))
-
-        result = gocardless_service.handle_incoming_webhook(
-          organization_id: organization.id,
-          body: events.to_json,
-          signature: 'signature'
-        )
-
-        aggregate_failures do
-          expect(result).not_to be_success
-          expect(result.error).to be_a(BaseService::ServiceFailure)
-          expect(result.error.code).to eq('webhook_error')
-          expect(result.error.error_message).to eq('Invalid signature')
-        end
-      end
-    end
-  end
-
-  describe '.handle_event' do
-    let(:payment_service) { instance_double(Invoices::Payments::GocardlessService) }
-    let(:service_result) { BaseService::Result.new }
-
-    before do
-      allow(Invoices::Payments::GocardlessService).to receive(:new)
-        .and_return(payment_service)
-      allow(payment_service).to receive(:update_payment_status)
-        .and_return(service_result)
-    end
-
-    context 'when succeeded payment event' do
-      let(:events) do
-        path = Rails.root.join('spec/fixtures/gocardless/events.json')
-        File.read(path)
-      end
-
-      it 'routes the event to an other service' do
-        gocardless_service.handle_event(events_json: events)
-
-        expect(Invoices::Payments::GocardlessService).to have_received(:new)
-        expect(payment_service).to have_received(:update_payment_status)
-      end
-    end
-
-    context "when event metadata contains payable_type PaymentRequest" do
-      let(:payment_service) { instance_double(PaymentRequests::Payments::GocardlessService) }
-      let(:service_result) { BaseService::Result.new }
-
-      let(:events) do
-        path = Rails.root.join("spec/fixtures/gocardless/events_payment_request.json")
-        File.read(path)
-      end
-
-      before do
-        allow(PaymentRequests::Payments::GocardlessService).to receive(:new)
-          .and_return(payment_service)
-        allow(payment_service).to receive(:update_payment_status)
-          .and_return(service_result)
-      end
-
-      it "routes the event to an other service" do
-        gocardless_service.handle_event(events_json: events)
-
-        expect(PaymentRequests::Payments::GocardlessService).to have_received(:new)
-        expect(payment_service).to have_received(:update_payment_status)
-      end
-    end
-
-    context "when event metadata contains invalid payable_type" do
-      let(:events) do
-        path = Rails.root.join("spec/fixtures/gocardless/events_invalid_payable_type.json")
-        File.read(path)
-      end
-
-      it "routes the event to an other service" do
-        expect {
-          gocardless_service.handle_event(events_json: events)
-        }.to raise_error(NameError, "Invalid lago_payable_type: InvalidPayableTypeName")
-      end
-    end
-
-    context 'when succeeded refund event' do
-      let(:refund_service) { instance_double(CreditNotes::Refunds::GocardlessService) }
-      let(:events) do
-        path = Rails.root.join('spec/fixtures/gocardless/events_refund.json')
-        File.read(path)
-      end
-
-      before do
-        allow(CreditNotes::Refunds::GocardlessService).to receive(:new)
-          .and_return(refund_service)
-        allow(refund_service).to receive(:update_status)
-          .and_return(service_result)
-      end
-
-      it 'routes the event to an other service' do
-        gocardless_service.handle_event(events_json: events)
-
-        expect(CreditNotes::Refunds::GocardlessService).to have_received(:new)
-        expect(refund_service).to have_received(:update_status)
-      end
-    end
-  end
 end
